move GetTypeInfo
