+++ ydb/core/tx/schemeshard/schemeshard_info_types.cpp
@@ -2531,34 +2531,5 @@ std::optional<std::pair<i64, i64>> ValidateSequenceType(const TString& sequenceN
     return {{dataTypeMinValue, dataTypeMaxValue}};
 }
 
-bool GetTypeInfo(const NScheme::IType* type, const NKikimrProto::TTypeInfo& typeInfoProto, const TStringBuf& typeName, const TStringBuf& columnName, NScheme::TTypeInfo &typeInfo, ::TString& errorStr) {
-    if (type) {
-        // Only allow YQL types
-        if (!NScheme::NTypeIds::IsYqlType(type->GetTypeId())) {
-            errorStr = Sprintf("Type '%s' specified for column '%s' is no longer supported", typeName.data(), columnName.data());
-            return false;
-        }
-        typeInfo = NScheme::TTypeInfo(type->GetTypeId());
-        return true;
-    } else if (const auto decimalType = NScheme::TDecimalType::ParseTypeName(typeName)) {
-        if (typeInfoProto.HasDecimalPrecision() && typeInfoProto.HasDecimalScale()) {
-            typeInfo = NScheme::TypeInfoFromProto(NScheme::NTypeIds::Decimal, typeInfoProto);
-        } else {
-            typeInfo = *decimalType;
-        }
-        return true;
-    } else if (const auto pgTypeDesc = NPg::TypeDescFromPgTypeName(typeName)) {
-        if (typeInfoProto.HasPgTypeId()) {
-            typeInfo = NScheme::TypeInfoFromProto(NScheme::NTypeIds::Pg, typeInfoProto);
-        } else {
-            typeInfo = pgTypeDesc;
-        }
-        return true;
-    } else {
-        errorStr = Sprintf("Type '%s' specified for column '%s' is not supported by storage", typeName.data(), columnName.data());
-        return false;
-    }
-}
-
 } // namespace NSchemeShard
 } // namespace NKikimr
