+++ ydb/core/tx/schemeshard/schemeshard_utils.cpp
@@ -554,7 +554,7 @@ bool ExtractTypes(const NKikimrSchemeOp::TTableDescription& baseTableDescr, TCol
         auto typeName = NMiniKQL::AdaptLegacyYqlType(column.GetType());
 
         NScheme::TTypeInfo typeInfo;
-        if (!NSchemeShard::GetTypeInfo(typeRegistry->GetType(typeName), column.GetTypeInfo(), typeName, columnName, typeInfo, explain)) {
+        if (!GetTypeInfo(typeRegistry->GetType(typeName), column.GetTypeInfo(), typeName, columnName, typeInfo, explain)) {
             return false; 
         }
